+++ packages/web3/src/index.ts
@@ -333,9 +333,9 @@ export default Web3;
  * Named exports for all objects which are the default-exported-object in their packages
  */
 export { Web3 };
-export { Web3Context, Web3PluginBase, Web3EthPluginBase } from 'web3-core';
+export { Web3Context, Web3PluginBase, Web3EthPluginBase, Web3PromiEvent } from 'web3-core';
 export { Web3Eth } from 'web3-eth';
-export { Contract } from 'web3-eth-contract';
+export { Contract, ContractDeploySend, ContractMethodSend } from 'web3-eth-contract';
 export { Iban } from 'web3-eth-iban';
 export { Personal } from 'web3-eth-personal';
 export { Net } from 'web3-net';